Allow standalone parsing of commands
